Don't ping local addresses on controller
When verifying addresses we specify the interface that we expect
the traffic to use.  For addresses that belong to the active
controller, this means that the connectivity test will fail because
the valid route is via the "lo" loopback device rather than the
external network interface.

Since the relevent addresses are statically configured anyway,
the simplest solution is just to skip checking for connectivity
on controller nodes.

Note: the change is logically just adding two lines, but the indentation
changes make it look more complicated.

Change-Id: I326ae03c5dc368692e43aeb7d5ba2ae95c957145