Remove tiller docker image in helm manifest
Previously the tiller image was included in the build and the helm
manifest loaded that image. The tiller image is no longer included in
the build. Without the tiller image present the helm manifest fails
during config_controller.

This changes the helm manifest to pull the docker image from its
external home.

Change-Id: If6ea726c0439e3a5e88fb7218c26b702a5f1122c
Signed-off-by: David Sullivan <david.sullivan@windriver.com>
Story: 2003491
Task: 24768