Detect BONDING_OPTS when applying network config
This commit adds BONDING_OPTS to the list of attributes to check for
changes when applying network config.

This commit also adds BONDING_OPTS as a special case inside of
normalized_cfg_attr_value.  This is necessary becase The BONDING_OPTS
attribute contains '=' characters, so is not correctly parsed by
splitting on '=' as done for other attributes.  Because this is only
being used to check for changes to BONDING_OPTS, rather than actually
using the returned value, this will return the whole line for
BONDING_OPTS.

Closes-Bug: 1789444
Change-Id: I01d1ef2bc80093d41fe95b68e4abdc309b574d47
Signed-off-by: Joseph Richard <Joseph.Richard@windriver.com>