Configure kubelet to ignore hugepages
By default, kubernetes assumes any hugepages on the host are
available for it to use, but it has the limitation that it can
only handle a single hugepage size per host.  This makes it less
useful for us, since we want to give vswitch 1GB pages but prefer
2MB pages for OpenStack VMs.

For now, the workaround is to tell kubernetes to ignore hugepages.

This also fixes up kubeadm.yaml.erb to use the new format for
kubernetes 1.12.
Change-Id: I0917781db4f1961d433399e0a9799f3a6253ca11
Story: 2003909
Task:  27085
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>