Mtce: Change SM Port scope to handle AIO config.
The mtc.ini file is updated a second time in AIO config.
Due to the scope of the SM ports being for controller only
and no defaults we see the sm port assignments missing in
AIO configs.

This update defaults the SM port numbers and changes the scope
of the parameters so that they get set on all node types for
all system types.

Testing included provisioning an AIO system.

Change-Id: Ib53921c4b59a9e67ed136a03504bdf0775de6dff
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>