Ceph for standard: System Inventory CLI, API, DB
In order to enable Openstack's helm charts on StarlingX we need a
distributed persistent storage for Kubernetes that leverages our
existing storage configurations. For this stage we will enable CEPH's
RBD to work with Kubernetes RBD provisioner through a new Helm chart.

Since RBD will be the persistent storage solution, CEPH support has to
be extended to the 1, 2 node and Standard configurations.

This commit enables CEPH on a standard configuration without dedicated
storage nodes by allowing users to add the 3rd Ceph monitor to a worker
node. It implements the functional part of this feature.

Details:
* Improve Ceph Monitor CLI to allow adding and deleting Ceph monitors
* Refactor Ceph Monitor API handling code;
* Add status and task DB column for ceph_mon table;
* Apply puppet manifests when a Ceph monitor is added through CLI or API
  at runtime on all nodes;
* Monitor manifest application status and report back puppet apply
  status on controllers;
* Delete monitor when worker is deleted;
* Add monitor for storage nodes;
* Minimal semantics.

Change-Id: Ie316bb611a006bbbc92ac22c52c3973cc9f15109
Co-Authored-By: Ovidiu Poncea <ovidiu.poncea@windriver.com>
Implements: containerization-2002844-CEPH-persistent-storage-backend-for-Kubernetes
Story: 2002844
Task: 28723
Signed-off-by: Ovidiu Poncea <Ovidiu.Poncea@windriver.com>