Set proper permissions on docker filesystem
The puppet manifests are currently setting the file permissions
for the docker filesystem (/var/lib/docker) to 750. However,
docker changes the permissions to 711. The result of this is that
whenever the puppet manifests are applied, the permissions are
changed, which has a cascading effect, causing docker to be
restarted and all running pods to be stopped. In the AIO-SX,
this happens when the compute manifests are applied, resulting
in all pods being torn down while kubernetes is in the middle
of trying to start them. This is bad.

The solution is to update puppet to set the permissions for
the docker filesystem to 711. That way, puppet will not
change the permissions or restart docker.

Change-Id: I0a7ca6f21947841afe4ca3a8ba6b4f49e58d5e8a
Story: 2004520
Task: 28573
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>