Ignore error on k8s taint removal from puppet
There are cases where the kubernetes taint is not present on,
or has already been removed from, a newly configured standby
controller. This causes the taint removal command run by the
puppet manifest to fail. This failure can be safely ignored,
so the command is updated by this commit to always return
success.

Change-Id: Icdb55738e052c65a28e44582e345038b0de83c37
Closes-Bug: 1815795
Signed-off-by: Don Penney <don.penney@windriver.com>