Fix to not restart alias interface
The alias interface may at times be brought up before its parent
interface (ethernet or vlan) depending on how the interface config
files are listed in the filesystem.
The fix is to not restart the alias interface. If there is a
change in the alias interface configuration, the parent interface
would be restarted instead.

Partial-Bug: 1817593

Change-Id: I83b88e1587c6468d85e14a50de934908586cbc9b
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>