Remove optionality of kubernetes from manifests
Any place where the kubernetes::params::enabled flag
was checked, it is always considered to be true.

This leads to some manifest code never being invoked
so a subsequent commit will clean up those code blocks.

The openstack manifests are not included in this change
because most of those will be removed in a later commit.

Story: 2004764
Task: 29821
Change-Id: Id47c69d6da2d243c607bcb3db0836073321e4146
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>