Enable Distributed Cloud configuration
The following changes are required to enable system
controller and sub cloud configuration in a distributed
cloud environment:

* Remove references to os-keystone-region-name as the
openstack patches that support it, have been removed.
* Change the iptables rule for the NAT entry, to only
apply, if the selected outgoing interface is the
OAM interface.
* Configure keystone endpoints, before configuring
openrc on subclouds
* Remove all openstack services, and users from the region
config and update the tox
* Disable nova, cinder and neutron api proxy

Only tested distributed cloud configuration as multi-region
configuration is not supported in the current release.

Story: 2004766
Task: 30017

Change-Id: I5c43e2112f34225aa9e23ff777c5333ae77efcdc
Signed-off-by: Tao Liu <tao.liu@windriver.com>