Fix application-apply of stx-openstack on simplex
The application-apply of the stx-openstack application on
simplex configurations has been failing since the barbican
chart was added to the application. The failure was due
to lost node status messages from the kubelet to the
kube-apiserver, which causes the node to be marked
NotReady and endpoints to be removed.

The root cause is the kubernetes bug here:
https://github.com/kubernetes/kubernetes/issues/74412

In short, the addition of the barbican chart added enough
new secrets/configmaps that the kubelet hit the limit of
http2-max-streams-per-connection. As done upstream, the
fix is to change the following kubelet config:
configMapAndSecretChangeDetectionStrategy (from Watch to
Cache).

Change-Id: Ic816a91984c4fb82546e4f43b5c83061222c7d05
Closes-bug: 1820928
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>