Disable PHC sanity check in case of software PTP mode.
boundary_clock_jbod performs a sanity check to make sure
that all of the ports share the same hardware clock device.
This option is not needed in case of software PTP mode.
Moreover it interferes with normal PTP operation in this
case and causes PTP clocks instability in a network.

Also, cleaning up unused pmon scripts for ptp4l and phc2sys
and adding services dependencies from mainline linuxptp.

Change-Id: If4bbe6af600dbdf38d301deafb7dc050a7754cad
Closes-bug: 1824218
Signed-off-by: Alex Kozyrev <alex.kozyrev@windriver.com>