Add support for multus and sriov CNI, device plugins
This commit adds support for the Multus, SRIOV CNI, and SRIOV
device plugins plugins.

It is intended that these would be used outside of the stx-openstack
application, to pass through devices for a user defined container.

The Multus CNI plugin enables attaching multiple network interfaces to
pods in Kubernetes

The SRIOV CNI plugin allows passing through of an SRIOV VF to a
container, with support for IP, vlan, DPDK  configuration.

The SRIOV device plugin is a kubernetes device plugin for discovering
and advertising SRIOV VFs in a Kubernetes host.

Sysinv changes have been made to populate the SRIOV device plugin
configuration file with appropriate values based on the host's
configured PCI-PT and SRIOV interfaces.

To enable SRIOV interfaces in a container, the user would create
a custom multus NetworkAttachmentDefinition with type 'sriov'
and refer to it in the Pod's k8s.vi.cni.cncf.io/networks annotation.

A few sample NetworkAttachmentDefinition and pod deployments can be found
here:

https://github.com/intel/sriov-network-device-plugin

Story: 2005208
Task: 29983
Task: 30054

Co-Authored-By: Litao Gao <litao.gao@windriver.com>
Co-Authored-By: Jackie Huang <jackie.huang@windriver.com>

Change-Id: I01587af6a4dc40fe7542b140e5bbee539d93740f
Signed-off-by: Steven Webster <steven.webster@windriver.com>