Enable platform helm repo for RPM installed charts
Add a new helm repository, 'stx-platform', designed to hold charts that
need to be delivered as part of the basic platform. These charts will be
installed via RPMs as part of install and patching.

Update the existing stx-openstack armada application manifests to
reference the new location of the existing 'starlingx' repo. The
'starlingx' repo will be renamed with a future commit to 'stx-apps'.

Enable multiple repository support when generating helm overrides for
the chart location.

This updates both the puppet manifests and ansible playbook for initial
and subsequent configuration scenarios.

Change-Id: I0caaa878a6c6781d038b48b8caa2aa507ee9568a
Depends-On: I4b1a3615a6bd5d0bdd834a1cdf27c05d5a1057a0
Depends-On: I096d5ac126efc97f9a0a0f54f1e02323d936281c
Story: 2005424
Task: 30644
Signed-off-by: Robert Church <robert.church@windriver.com>