fix two coredns pod in simplex system
There are 2 coredns pod in default. For simplex system, use kubectl
scale cmd to set the replicas to be 1.
A coredns class is created for coredns configuration. And is configured
to run as initial configure only, the same as kubernetes::master::init.

Test:
Pass simplex and duplex deploy test, and coredns pod is 1 for simplex
and 2 for duplex. affinity and nodeselector setting for duplex is kept.

Closes-Bug: 1822880

Change-Id: Ic2bbdc614acc0a04f91bec71691c2ff097218915
Signed-off-by: Shuicheng Lin <shuicheng.lin@intel.com>