Add Commands For Docker Registry Image Cleanup
Add commands for
registry-image-list, to list the existing images in the repo
registry-image-tags, to list the tags of a specified image
registry-image-delete, to delete a tag of a specified image
registry-garbage-collect, to delete space on the filesystem from
images that do not have any tags referencing them

It is currently impossible to delete images from the local Docker
registry on the controllers. This can be an issue if images are
continuously pushed. The only solution was to grow the file system.
This commit provides commands for a user to find unwanted images and
to delete them from the controller local Docker registry in order to
free up space on the file system.

There is a possibility of the registry getting stuck in read-only mode
due to garbage collect if puppet dies or there is a swact while
garbage collect is running. In this situation, the user should run
garbage collect again. Getting stuck in read-only mode does not
prevent image pulls or kubernetes deployments from existing images
from starting or completing.

Story: 2002840
Task: 28621

Change-Id: I4bb9301a1165db8b860418c413aa3238169bab03
Signed-off-by: Jerry Sun <jerry.sun@windriver.com>