Provision/deprovision dbmon dynamically
This change is to provision and deprovision dbmon service on an
AIO/DX or distributed cloud systemcontroller after stx-openstack
application is active.

1. Add active state to kube-app.
A kube-app is active after the first time it is applied successfully,
and deactive when it is being removed.
An util function is provided to access the active state of a given
kube-app.

2. apply stx_openstack runtime manifest to provision/deprovision
based on the active state of stx-openstack application

Story: 2005486
Task: 30922

Change-Id: Icbe63de21f27d7a804715ec72a04d1aba3910b09
Signed-off-by: Bin Qian <bin.qian@windriver.com>