Disable resource containers for kube-proxy
This disables resource containers for kube-proxy since that feature
is not needed and creates subsequent issues when the cpu manager
static policy is enabled.

We would see see a flood of cgroup device and cpuset related error
logs when the cpu manager tries to update cpusets for the nested
cgroup. We would also see that kube-proxy could not update its cpuset
to the DefaultCPUSet as intended. With resource containers disabled,
an unneeded directory level in the cgroup hierarchy is removed.

The same issues with resource containers are reported here:
https://github.com/kubernetes/kubernetes/issues/66614
https://github.com/kubernetes/kubernetes/issues/54967
https://github.com/kubernetes/kubernetes/issues/17619
https://github.com/kubernetes-sigs/kubespray/issues/3518

Note that specifying the --resource-container option is deprecated,
and the setting defaults to "/kube-proxy" when empty string is
provided. Configuring the value to "/" effectively disables this
feature, where kube-proxy will just log a non-fatal error message.

Change-Id: Id73303afb9b5aa8fa628cc17e03f3457787cfa22
Story: 2005192
Task: 33484
Signed-off-by: Jim Gauld <james.gauld@windriver.com>