Restart kubelet after hugepage allocation
On AIO systems, kubelet is launched prior to the application
of the worker puppet manifest. However, in some cases, the
hugepage allocation is occurring as part of this manifest
application, and kubelet does not have these hugepages in
its availability.

To address this, the kubernetes manifest is updated to define
a refresh dependency against the hugepage allocation on AIO,
in order to restart kubelet if needed.

Change-Id: I3289b83fb5bf93553b786694a153be84a97ebe8d
Closes-Bug: 1830927
Signed-off-by: Don Penney <don.penney@windriver.com>