AIO-SX: Fix adding OSD at runtime hanging when /dev/rbd* are mounted
This mostly happens when stx-openstack application is running
as there are a couple of pods that uses RBD PVCs.

The reason of the hang is that, on AIO-SX, all OSDs are stopped
while new ones are configured. This leads to data on the /dev/rbd*
to be inaccessible when runtime puppet manifests executes
'ceph-disk list' to query the state of all storage devices in
the system.

This commit removes the code that stops existing OSDs on all
deployments. A side effect is a decrease in time needed to
add the OSDs on a node with many OSDs already configured.
For e.g. time for adding an OSD on a system with 7 OSDs can
take more that 2 minutes, while with this change it drops to
around 30 seconds.

Change-Id: I41a65d42a45a90b00f65509f4e536eb1c345a91b
Closes-Bug: 1829855
Signed-off-by: Ovidiu Poncea <ovidiu.poncea@windriver.com>