Fix the helm-upload failure on controller-1
The user "www" is used to upload charts by using
helm-upload cmd. The root cause for the helm-upload
failure is the uid and gid of user "www" do not match
the uid and gid on controller-0.

Currently, www is created on controller-0 during ansible
configuration and the uid is set to "1877" explicitly.
On controller-1, it's created in puppet manifest but
it doesn't set the uid.

The user "www" should really be added in setup package so
that it's created after each node installation and it makes
sure the uid and gid are consistent on each node.
This is done in review https://review.opendev.org/#/c/668034/

However, we want to add the uid and gid in puppet as well
to ensure they are identical on both controllers in case
the user is not created during setup rpm installation.

Change-Id: I54035739c15eb0f5254f00f5c22d44925316e81c
Partial-Bug: 1833603
Signed-off-by: Angie Wang <angie.wang@windriver.com>