Support Single huge page size for openstack worker node
Kubernetes only supports a single huge page size per worker
node. Prior to kubernetes 1.15, the huge page feature could
be disabled via a feature gate. In kubernetes 1.15, the
feature gate has been removed so huge page support is always
on in k8s.

This update removes the conditional disabling of the hugepage
feature and enforces the provisioning of a single page size
per worker.

When vswitch type is set to ovs-dpdk or avs, the application
huge pages size goes with the vswitch huge pages size.

This update also changes the auto-provisioning of VM huge
pages to 1G as there is no auto-provisioning in virtual
environment.

Story: 2006295
Task: 36006

Change-Id: I84d4959b420584fdcdf8a8664a6f4855c08ec989
Signed-off-by: Tao Liu <tao.liu@windriver.com>