ceph: explicitly bind ceph-mgr to controller address
ceph-mgr has no corresponding sections in ceph.conf
When the service starts it binds and advertise a
local address.

In the case of #1843082 ceph-mgr on controller-0 binds
to the floating management IP address [face::5]. When
controller swacts ceph-client API requests are sent
to [face::5] but there is no service listening on that
endpoint and host-lock command hangs.

Fix: use ceph.conf to explicitly bind ceph-mgr to a
host-specific controller address.

Change-Id: I811bfa7410770eb91eb4519cfab3e6febf381e6d
Closes-bug: 1843082
Signed-off-by: Daniel Badea <daniel.badea@windriver.com>