Change the page size mounted at /dev/hugepages
The libvirt helm chart assumes that the default hugepage
size is mounted at /dev/hugepages.

The default size mounted at /dev/hugepages was fixed to 2M,
while the kernel default huge page size was set to 1G on
Baremetal, as a result the libvert pod failed to start since
it was unable to write to the hugepage mount.

This update changes the page size mounted at /dev/hugepages
to be the same as the kernel default hugepage size.

Change-Id: Icc0326b99338ca7c06b113e6991f01b838030aca
Closes-Bug: 1844576
Signed-off-by: Tao Liu <tao.liu@windriver.com>