Change group of /etc/ceph/ceph.conf file
The sysinv-api is currently running as the sysinv user, but with
the group set to root. This is not correct - the sysinv-api should
run with the normal groups for the sysinv user, which include
sys_protected.

However, the sysinv-api needs access to the /etc/ceph/ceph.conf
file (to use the ceph API). To allow this, change the group
of the ceph.conf file to sys_protected, which will allow the
sysinv-api to access the file.

Change-Id: I422fada2c5411d84f5542722fceea8c64e3ff6c7
Story: 2006590
Task: 36724
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>