Removing service catalog insertion from dcorch proxy
requests going through the dcorch proxy were having the entire service
catalog tacked on during the authtoken filter stage, this was resulting
in the header size growing too large for sysinv to handle the forwarded
requests.

This commit sets keystone_authtoken/include_service_catalog to False in
the dcorch settings to prevent this.

Tested by installing a subcloud, bringing online and managing, as well
as doing sysinv queries to SystemController.  I've tested with 200
subclouds in dcmanager without issue.

Change-Id: Ic47c062bd8b5376084d27a9378c131650d9ec2da
Closes-Bug: 1856740
Signed-off-by: Tyler Smith <tyler.smith@windriver.com>