Configure keystone http timeouts for dcmanager
The dcmanager currently has no http_connect_timeout set for
keystone connections. That can result in an attempt to contact
keystone (e.g. to get a token) taking several minutes to
timeout if the keystone api is not reachable (e.g. if a subcloud
is powered down).

Changing the http_connect_timeout to 10s and configuring
http_request_max_retries as 3 (that is also the default but
adding this to the puppet module allows for easy changes in the
future).

Change-Id: I6a62846e7f4e75e9b2f0705f59818243ea909e41
Partial-Bug: 1854894
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>