lowlat: enable ktimer_lockless_check if it exists
Enable check for raising timer interrupt only if one is pending.
This allows nohz full mode to operate properly on isolated cores.
Without it, ktimersoftd interferes with only one job being
on the run queue on that core, causing it to drop out of nohz.

If ktimer_lockless_check doesn't exist in the kernel, then no
error is reported ie. it just fails silently.

Closes-Bug: 1870456
Change-Id: I93d0fab3e9f4f56f9afb9bbfaa04882cf9068db5
Signed-off-by: Jim Somerville <Jim.Somerville@windriver.com>