Allow k8s upgrades to any release if necessary
The default behaviour of the "kubeadm upgrade apply" command is
to only allow upgrades to stable kubernetes versions. However,
for both testing purposes and for potential critical fixes in
the future, it may be necessary to upgrade to a release
candidate or other release that kubernetes deems as unstable.
Adding in the appropriate options when calling the "kubeadm
upgrade apply" command to make this possible.

Change-Id: I164caf495ee3680f549d651b97e7e502b1172c70
Story: 2006781
Task: 37578
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>