Support for FPGA FEC devices
This commit introduces puppet support for forward error correction
(FEC) devices that are enabled on an Intel N3000 FPGA device.

The FEC device is intended to be used with a DPDK application,
with the device itself being bound to the igb_uio driver. As
such, support has been added for enabling a number of SR-IOV
virtual functions on the PF and binding both the PF and VF to
a supported driver.

Story: 2006740
Task: 39948

Depends-On: https://review.opendev.org/733724
Change-Id: If85f9ac93c3d8c2418c7128f2d3ef5fd4a4509fd
Signed-off-by: Steven Webster <steven.webster@windriver.com>