Run docker login at worker startup
The N3000 FPGA tools run in a docker container on the worker nodes including
the standby controller in an AIO-DX environment where it's difficult to
obtain the docker password.

As such, let's have the root user run "docker login" for the local registry
during early startup. That will then ensure that the sysinv FPGA agent can
pull docker images from the local registry in order to talk to the N3000
FPGA.

Change-Id: I6d1f2c1c20b814c4c91cb033fc9f36261d8686d3
Story: 2006740
Task: 39931
Depends-On: Ie9eea8b670f0c82f0108ccbec978500e1bc0eb6a