Remove annotation before oam if gnp re-apply
When service parameter http_port is modified and applied,
platform::firewall::runtime is triggered and oam if gnp is re-applied.
But the resourceVersion in existing gnp's annotation will require the
re-apply to provide matching resourceVersion in the yaml file, and will
fail the re-apply if it doesn't. This update fixes this issue by
removing annotation before the gnp is re-apply.

Change-Id: Idc373c206cf2dab5e02d98e11e1c9b18a605dc37
Closes-Bug: 1873600
Signed-off-by: Andy Ning <andy.ning@windriver.com>