Run kubeadm command with specific configuration file
Puppet manifests use the kubeadm command during initializing
kubernetes nodes. However, it is not passing the --kubeconfig
parameter to this command, so the command will search for the
kubeconfig file. Normally it finds this file in
/etc/kubernetes/admin.conf, but if a user accidentally (or
intentionally) creates a config file at /root/.kube/config, this
file can take precedence and cause the kubeadm command to fail.

This commit updated the command with /etc/kubernetes/admin.conf as
its configuration file explicitly.

Change-Id: I2942fa4a275145cc3a1b6c6fdb0f3827a244f1bb
Closes-Bug: 1882678
Signed-off-by: Andy Ning <andy.ning@windriver.com>