Handle N3000 FPGA reset
A recent config commit (6d639fca5d) was introduced that resets any
Intel N3000 FPGA devices on startup.  After an FPGA reset, it is
possible that the PCI addresses of the underlying FEC device may
change, in addition to the resetting of any configured drivers and
SR-IOV VF configuration on the FPGA FEC and NIC devices.

To handle the reset, it is required to do the following:

- Do not allow a user to configure an FEC device before the first
unlock (to allow the reset to occur).

- Delay the configuration of FEC and NIC devices until the reset
  has occurred.

Depends-On: https://review.opendev.org/737388
Story: 2006740
Task: 39948

Change-Id: I6d1bd988192dca868b8cc972045933ee8ac75e7c
Signed-off-by: Steven Webster <steven.webster@windriver.com>