Remove hardcoded /scratch size from manifest
This update removes the hardcoded sizing of /scratch on controllers
from the filesystem manifest, and instead using the configured lv_size
from sysinv. Currently, on a reinstall of the controller, the
configured size would be ignored, and /scratch would be resized to the
hardcoded 16G value. Instead, the initial sizing is updated to 16G in
the controller kickstarts, sizing /scratch to the desired default size
on installation. If the filesystem has been resized in sysinv
configuration to a higher value, then the filesystem manifest will now
resize accordingly on a reinstall.

Depends-On: https://review.opendev.org/738084
Change-Id: Id0f7072d555018f8b19b91fd291ac3b008c10173
Closes-Bug: 1885168
Signed-off-by: Don Penney <don.penney@windriver.com>