Fix SR-IOV enable for VF type interfaces
The puppet config for type VF interfaces deliberately has the
number of virtual functions set to None, as they borrow from
their parent SR-IOV interface. That is, it's only necessary to
set the number of VFs on the parent interface.

A previous commit changed the check on the num_vfs parameter to
ensure it is > 1.  We need to also check that this parameter is
defined.

Closes-Bug: #1885173
Change-Id: I671af0bc16aed3a38d2ebe84b0f6b7195180e97f
Signed-off-by: Steven Webster <steven.webster@windriver.com>