Enable SR-IOV device plugin restart
In an AIO system, it is possible for the kube-system pods, including
the SR-IOV device plugin to start before the worker manifest finishes
enabling and binding drivers to network interface VFs.

Since the device plugin does not periodically (re)scan the PCI bus,
it is required to restart the plugin after completing the SR-IOV
driver bind to ensure that the full allocatable set of VFs is
inventoried.

Note that this can probably be mitigated in the future when the
device plugin is converted to use helm / config map rather than
having puppet write it's /etc/sriovdp/config.json file.

Change-Id: I7972d7a56c2d38884238f7c7818892d0a5b33a0e
Closes-Bug: #1885229
Signed-off-by: Steven Webster <steven.webster@windriver.com>