Enable subcloud keystone public endpoint access on OAM IF
Currently keystone public endpoint on subcloud is not accessible from
OAM IF for two reasons:

- haproxy is not configured for keystone
- The OAM global network policy doesn't allow keystone public endpoint
  port (5000) to go through.

Inaccessible to keystone causes remote CLI not working as keystone is
the first service the CLI needs to access to authenticate the requests.

The change enabled the access by updating haproxy configuration and
the OAM gnp.

Change-Id: I540b44994c8c20f13d14d2e26eda6351b0f916e7
Closes-Bug: 1886712
Signed-off-by: Andy Ning <andy.ning@windriver.com>