Database connection exhausts during dcorch audit
In a distributed cloud system with lots of subclouds, seeing QueuePool
limit exception during dcorch audit, and the particular subcloud cannot
synced with the system controller until the next audit attempt. This
change increases the overflow limit to 500 on the db connection and
restrict the db connection size to 1 to prevent the idle connection
remaining in the pool.

Change-Id: I93eae7b7e8927f4a70d0073bf4619bcf0de15df7
Closes-Bug: 1889292
Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>