Update drbd resize operation to run after puppet manifest
There is a potential race condition in triggering the filesystem
resize check.  The resize check is triggered on the start of
a controller filesystem drbd resize operation, however, if another
configuration is applied before the resize check may be early,
in which case the lv would be resized but not drbd.

In the drbd manifest update to trigger the drbd filesystem resize
check after the lv filesystem  have been resized.

Change-Id: I97038d0cc9d57ab639f85c2282677364b44249c0
Closes-Bug: 1894003
Signed-off-by: John Kung <john.kung@windriver.com>