Fix route config handling for DOR
In a DOR (Dead Office Recovery, where all nodes reboot at once), both
controllers are rebooting at the same time. This means that there is
no active controller from which to retrieve puppet data in order to
apply the controller manifests. As such, we have to be careful not to
rely on having the controller manifests run on every controller boot,
for things like launching services or any sort of changes in a
volatile file system like /var/run.

The route configuration optimization changes that were added in
https://review.opendev.org/703034 inadvertently relied on existing
network configuration data being cached in /var/run, however. As a
result, a route configuration change after a DOR of a system with
standard controllers would end up running with no cached network
config data (an AIO system would have this data generated as part of
applying the worker manifest), and the apply_network_config.sh utility
would think that all network interfaces have been removed from the
system. It would then proceed to apply that config, deleting the
interfaces and taking down all networking.

This commit enhances the apply_network_config.sh to introduce a
--routes option to separate the route configuration operations from
the rest of the networking config. When a route is added or deleted,
then, only the route config changes are processed, ignoring network
interfaces.

Additionally, this adds a check in the interface section of the
apply_network_config.sh utility to verify that at least 'lo' exists.
Since the loopback interface should always exist, its absence would
indicate that the interface config data is missing or corrupted, and
is unsafe to apply.

Change-Id: I5583ec916aee8117e5686cfb10fb18ddda4806b1
Closes-Bug: 1895693
Signed-off-by: Don Penney <don.penney@windriver.com>