Remove admin_token from keystone config
Currently the admin_token is still set with a value in keystone.conf
though it is disabled after bootstrap and no longer in use. This update
removes it during controllers unlocking as a security enhancement.

This update also fixes an ceph issue that would be triggered by the
above change and cause ceph.pp not generating ceph.conf properly
due to a resource creation disorder.

Change-Id: I4093bca40fad3724e89d902aae36d26f85aebd60
Closes-Bug: 1900726
Signed-off-by: Andy Ning <andy.ning@windriver.com>