Update grub defaults from  manifests
This update adds a utility, puppet-update-default-grub.sh, that is
called from the grub.pp and compute.pp manifests when they call grubby
to update the grub cmdline. This ensures that the changes being made
via grubby will have corresponding changes reflected in the grub
defaults file, so that added options are not lost if the grub cmdline
is updated via calls to grub2-mkconfig.

Change-Id: Ibf74ddf9d0445e92dd8903c67e95901d856ad373
Closes-Bug: 1901559
Signed-off-by: Don Penney <don.penney@windriver.com>