Ensure sriovdp is deleted after dev bindings
This change replaces the daemonset rollout restart command with a more
spcific pod delete command that only runs if there is an
sriov-device-plugin pod present on the node. Using the pod delete
command ensures that an existing device-plugin pod is terminated before
the worker manifest completes. The rollout restart command did not
ensure that the pod was terminated before the manifest completed and
could allow user pods to be assigned incorrect VFs if they started up
before the device-plugin pod terminated.

This addresses an issue where pods restarted by k8s-pod-recovery could
be assigned to incorrect VFs if they were started while the
sriov-device-plugin was shutting down. Waiting for the device-plugin
to completely terminate before proceeding with pod-recovery ensures that
the device-plugin will have an accurate view of all device bindings and
can allocate VFs correctly.

Closes-Bug: 1900736

Change-Id: I30fd602208d14ac887d5417fd87f27f23050f670
Co-Authored-By: Steven Webster <steven.webster@windriver.com>
Signed-off-by: Cole Walker <cole.walker@windriver.com>