Make sure OSDs are placed on nodes
With the recent work that added the possibility to take a backup on
controller-1 and restore on controller-0 the [osd.X] entries are removed
from ceph.conf before controller-0 unlock.

An [osd.X] entry must be present in ceph.conf so that a ceph-osd process
is spawned and the directory structure for that osd is created in
/var/lib/ceph/osd/ceph-X.

Mtc runs an initialization script which expects that directory to be
populated, otherwise it fails and places the node in a degraded state.

For AIO systems the [osd] entry was present populated, but for STANDARD
it was not. This commit ensures the entry is populated for both AIO and
STANDARD systems.

Tested full install + platform B&R on AIO-SX, AIO-DX, STANDARD with
controller storage, all with ceph.

Closes-Bug: 1901636
Change-Id: I669ad3b22f59136253d42cdf6ac603fba31000be
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>