Update dnsmasq config for slow DNS servers
When a configured DNS server is taking a long time to respond to
unknown domains or hosts, registry interactions like push, pull,
and querying for images through system commands will fail due to
hostname resolution for registry.local. This is because it attempt
to resolve registry.local using the A record first, which times out
since it is hitting the configured external DNS server. This
prevents the process from looking up the AAAA record which would
resolve to the dnsmasq CNAME record. This commit updates the dnsmasq
config to prevent forwarding the local domain to upstream servers.

Change-Id: Ic3cf6aae87f8f2d5c61a24db00a4cb814c20aac6
Closes-Bug: 1904885
Signed-off-by: Jerry Sun <jerry.sun@windriver.com>