Fix directory permissions for /var/log/rabbitmq
Updated /var/log/rabbitmq directory permissions to 750 from 755
to disallow world access to rabbitmq log files but at the same
time to allow group access.
The changes are made to comply as much as possible with
openscap rules security requirements.
Verified that installation is successful for AIO-SX
and Standard 2+2 system configurations.

Story: 2008037
Task: 40694

Change-Id: I1c0112575033c04983c56298e2131882911333de
Signed-off-by: Carmen Rata <carmen.rata@windriver.com>