Reprovision SM services on duplex
Update SM provisioning for duplex to reprovision services
if needed. The default configuration in SM is duplex services,
and a simplex node will reprovision these to be simplex. In
order to support SX to DX migration, these services will also
be reprovisioned on duplex to ensure the configuration
is correct.

Story: 2008587
Task: 41743

Change-Id: Ifb61a6046c680d0dee7c76660397c6fe8c2cbe73
Signed-off-by: Don Penney <don.penney@windriver.com>