platform puppet: Config ACC100 bbdev with QMGR val
The ACC100 PF and VF configuration takes the same puppet
config code path as the N3000 except that the ACC100 does
not require a reset, but requires bbdev config.

This patch adds platform::devices::acc100::fec class to
exec pf-bb-config to configure QMGR on the Intel ACC100
(Mt. Bryce) with number of 5G UL/DL qgroups and configures
the device with the number of VF's.

Story: 2008440
Task: 41530

Depends-On: https://review.opendev.org/c/starlingx/integ/+/775252

Signed-off-by: Babak Sarashki <zbsarashki@gmail.com>
Change-Id: I7d42852009fedba5136d9d726092f273ef41c7fd