Use kubelet.conf instead of admin.conf on worker nodes during upgrade
Specifying a config file that does not exist causes kubelet upgrades
to fail on worker nodes when some of the commands return errors.

admin.conf does not exist on worker nodes, but exists on controllers.
The code has been updated to use kubelet.conf during worker kubelet
upgrade actions.

The worker init code has also been changed when pulling the pause
image so that it does not try to contact k8s.gcr.io.
The kubernetes-version needed to be passed in when querying for the
pause image.

Story: 2008137
Task: 41828
Change-Id: I6565132bd587927bd26c845c2ea56a995ac6da1c
Signed-off-by: albailey <Al.Bailey@windriver.com>