Increase timeout for sriovdp deletion
In an AIO-SX, pods get launched by kubelet soon after puppet is done
with controller's manifest but is still working with worker's manifest.
When pods are several the concurrency may lead the SRIOV device plugin
to not be deleted (then restarted) in the expected time frame.
The final solution for this problem is on the way, by refactoring the
current AIO to orchestrate between pods bring-up and worker setup. In
the meantime, the workaround solution in this change is the increase
of the original timeout for deletion of the SRIOV device plugin.

Closes-Bug: 1916620
Implements: increased timeout value to manage sriovdp in kubernetes.pp
Signed-off-by: Douglas Henrique Koerich <douglashenrique.koerich@windriver.com>
Change-Id: I0f6fb20a0ed5086fc80794b35715eea8d3d74cb8