puppet manifest apply check hieradata rsync
Update the puppet manifest apply to check whether the hieradata
has been rsync successfully.  Check return value and, in certain
cases reattempt, before continuing.  This is needed because
the hieradata is actually generated by the controller node,
and this script may be running on another host.

It has been observed that there are instances on worker host
whereby some of the hieradata is missing (e.g. missing
system.yaml openstack_host in puppet.log).

Verified:
    install, deployment and sanity on multinode and AIO
    backup and restore with Ceph
    platform upgrade

Change-Id: I9e7a0a02dd28c06d914fafe8234f4fee5e05247c
Closes-Bug: 1917229
Signed-off-by: John Kung <john.kung@windriver.com>