Increase min_free_kbytes to 256M for storage to avoid OOM issue
Help to prevent the OOM issue that it failed to allocate memory
with error message 'page allocation failure: order:2, mode:0x104020'

As the min_free_kbytes in the linux documentation shows:
This is used to force the Linux VM to keep a minimum number
of kilobytes free.  The VM uses this number to compute a
watermark[WMARK_MIN] value for each lowmem zone in the system.
Each lowmem zone gets a number of reserved free pages based
proportionally on its size.

Keeping more memory free in those zones means that the os itself is
less likely to run out of memory during high memory pressure and high
allocation events.

Based on the issue occurs on the storage node so far, we only update
the value for the storage node.

Closes-Bug: #1924209

Change-Id: Iae2e5a0787f69c62ba5da53663371fd2be148e15
Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>