Fix interface (re)configuration for single-nic system
Currently, the apply-network-config manifest step launches a script
that detects differences between puppet's view of what the
ifcfg-* network scripts should be and what the value
of the ifcfg files actually are in the /etc/sysconfig/network-scripts/
directory.

If there are differences, the puppet representation of the interface
configuration is copied to the system network-scripts directory and
the interface is brought down and up to apply the config.
If there are no changes between the puppet view and the system view,
the interface is left alone.

An issue can occur in a single-nic system comprising a physical
lower ethernet interface configured for SR-IOV with upper vlan
interfaces (oam, mgmt, etc).  If the lower interface is
re-configured, it is subsequently brought down/up to apply
the changes.  This causes the upper vlan interfaces to also
be brought down by the kernel.  In the case of an IPv6 system,
the interfaces will lose their addresses as well as any configured
default route.  In the case of an IPv4 system, the default route
will be wiped out, which could cause issues in a distributed cloud
environment.

This commit addresses the issue by detecting whether any lower
interface associated with a vlan interface has been marked for
re-configuration.  If this is the case, the vlan interface is
also added to the up/down list to cause it to re-apply the
existing static configuration (if it is not already in the list).

Closes-Bug: 1926366
Signed-off-by: Steven Webster <steven.webster@windriver.com>
Change-Id: I40177900ef58a9619fecb34ceffc412f31d1a965