On runtime, kube-sriov-device-plugin needs to be restarted
The previous correction for bug 1918139 removed the sriov plugin
restart necessary during runtime, done during the interface sriov
assign to a datanetwork (allowed on an unlocked AIO-SX). Without
it, the pod creation will not be able to use a datanetwork created
on runtime.

The correction bring back the platform::kubernetes::worker::sriovdp
class to be used only on runtime

Closes-Bug: 1918139

Signed-off-by: Andre Fernando Zanella Kantek <AndreFernandoZanella.Kantek@windriver.com>
Change-Id: Ied19bf3138b58b279b350d067ae0c1080e220f31