AIO-SX reboots after change OAM ip address
On HW tests, it was detected that openstack-endpoints restart was
happening at the same as the service-manager restart, this creating
a conflict that preventing SM services to reach enabled-active.
This was provoking the reboot.

The correction creates openstack::keystone::endpoint::runtime::post
class to be executed the post stage and not on the main stage, to
avoid conflict with service-manager

Also marking platform::network::runtime to be run at the pre stage
to avoid some encountered apply errors related to the delay of
haproxy bringup due to the lack of the IP address on the interface
as it was only configured later. This way the other restarted
services will have the address on the interface as restart happens

Tested on AIO-SX, by monitoring manifest apply and validating that
no reboot happens

Closes-Bug: 1927275

Signed-off-by: Andre Fernando Zanella Kantek <AndreFernandoZanella.Kantek@windriver.com>
Change-Id: Ia70a3395753e43b3c1e2c037818c8c23e4ec0fd6