Add SR-IOV rate-limit dependency
Currently, the binding of an SR-IOV virtual function (VF) to a
driver has a dependency on platform::networking.  This is needed
to ensure that SR-IOV is enabled (VFs created) before actually
doing the bind.

This dependency does not exist for configuring the VF rate-limits
however.  There is a chance that the VF rate-limiting configuration
happens before the VFs are actually created.

This commit fixes the issue by creating a dependency on
platform::networking from the sriov::config class, which ensures
the VFs are created before both driver binding and rate
limiting configuration occurs.

Closes-Bug: #1927758
Signed-off-by: Steven Webster <steven.webster@windriver.com>
Change-Id: Ic452247eb8c980e1b18bdc54832eb635d7a9fc54