Safe restart of the etcd SM service in etcd upgrade runtime class
While upgrading the central cloud of a DC system, activation failed
because there was an unexpected SWACT to controller-1. This was due
to the etcd upgrade script. Part of this script runs the etcd
manifest. This triggers a reload/restart of the etcd service. As this
is done outside of the sm, sm saw the process failure and triggered
the SWACT.

This commit modifies platform::etcd::upgrade::runtime puppet class
to do a safe restart of the etcd SM service and thus, solve the
issue.

Change-Id: I3381b6976114c77ee96028d7d96a00302ad865ec
Signed-off-by: Jessica Castelino <jessica.castelino@windriver.com>
Closes-Bug: 1928135