Create pmond ceph conf on all storage nodes
Currently the pmond ceph config file is only
created on hosts that run ceph-mon, which is not
the case for all storage nodes.

We need to configure pmond to watch ceph on all
storage nodes so that other ceph processes (like ceph-osd)
are monitored and restarted when necessary.

Tested by installing a Standard with dedicated
storage nodes and checking the presence of the
config file. I also killed the ceph-osd process
and verified that pmond restarted it.
Tested the runtime_osds class by doing a
system host-stor-add on a AIO-SX and verifying that
the pmond file was deleted and recreated.

Closes-Bug: #1929584
Signed-off-by: Isac Souza <IsacSacchi.Souza@windriver.com>
Change-Id: I8ec7f49ce98cf0cc40f9fe53e992f93ed4b2edaa