Enable N3000 FPGA reset if device is present
As part of the switch to the singular aio manifest, the reset of the
N3000 FPGA is no longer being performed at boot by the fpga agent.
Rather, it is now initiated on demand by the puppet manifest if an
N3000 device is configured for SR-IOV.  This can cause a problem on
the first configuration of the device as the PCI address of the FEC
device changes after reset.  That is, on first configuration, a user
will be modifying the device based on a PCI address that will not
exist after the device is reset.  This will cause the subsequent
enablement of SR-IOV VFs and binding to respective PF/VF drivers
by the puppet manifest to fail.

This commit fixes the issue (with the corresponding stx/config
change) by performing the reset of the device regardless of whether
it has been configured or not.

Since there is an existing semantic check to ensure that the device
can not be modified before the first unlock, the new check for
whether the device's PCI address actually exists on the system will
ensure that the device will be reset but not enabled for SR-IOV on
first unlock.  All subsequent modifications on the device will then
be done against the PCI address that exists after reset.

This commit also removes the check on the number of VFs being
greater than 0.  This allows the device to essentially be
'unconfigured' (number of VFs set to 0).  Rather, a check is done
to ensure that the configuration of the number of VFs is only
done if the current value for this parameter is different than
what is requested by the user.

Depends-on: https://review.opendev.org/c/starlingx/config/+/795123
Partial-Bug: #1931128
Signed-off-by: Steven Webster <steven.webster@windriver.com>
Change-Id: Iccb26afab30826935362a8c12b1d341b7b1015d0